deployment/backend/requirements.txt
flask
flask-cors
pymongo
boto3
tensorflow
Pillow